move config from utils namespace to top level
