+++ expertise/models/centroid/centroid.py
@@ -1,11 +1,11 @@
 import numpy as np
 
-# import fastText as ft
-
 import torch
 import torch.nn as nn
 from torch.nn import BCEWithLogitsLoss
+
 from expertise import utils
+from expertise.config import Config
 
 from expertise.evaluators.mean_avg_precision import eval_map
 from expertise.evaluators.hits_at_k import eval_hits_at_k
@@ -234,5 +234,3 @@ def eval_map_file(filename):
 def eval_hits_at_k_file(filename, k=2, oracle=False):
     list_of_list_of_labels,list_of_list_of_scores = utils.load_labels(filename)
     return eval_hits_at_k(list_of_list_of_labels, list_of_list_of_scores, k=k,oracle=oracle)
-
-
